one interface class
Only interface class is maintained for any api, we need to create only different class files
